Latihan-Flask/requirements.txt
Flask
requests
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
flask-bcrypt
flask-login
flask-dance
flask-restful
Flask-JWT
flask_cors
flask-jwt-extended
is-safe-url
python-dotenv
python-slugify
